new/requirements.txt
together>=1.0.0
anthropic>=0.30.0
python-dotenv>=1.0.0
